Make API functional again
